requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
antspynet=0.2.0=pypi_0
antspyx=0.3.4=pypi_0
apex=0.1=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.9.24=py38h06a4308_0
charset-normalizer=2.1.1=pypi_0
chart-studio=1.1.0=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
dm-tree=0.1.8=pypi_0
flatbuffers=23.1.21=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=pypi_0
idna=3.4=pypi_0
imageio=2.22.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intensity-normalization=2.1.4=pypi_0
joblib=1.2.0=pypi_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=15.0.6.1=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.2=pypi_0
ncurses=6.3=h5eee18b_3
networkx=2.8.8=pypi_0
nibabel=3.2.2=pypi_0
ninja=1.10.2=pypi_0
numpy=1.23.4=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.5.1=pypi_0
patsy=0.5.3=pypi_0
pillow=9.3.0=pypi_0
pip=22.2.2=py38h06a4308_0
plotly=5.11.0=pypi_0
progressbar=2.5=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
pyqt5=5.15.7=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=pypi_0
python=3.8.13=haa1d7c7_1
python-dateutil=2.8.2=pypi_0
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.3=pypi_0
rsa=4.9=pypi_0
scikit-fuzzy=0.4.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.39.3=h5082296_0
statsmodels=0.13.5=pypi_0
tenacity=8.1.0=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.30.0=pypi_0
tensorflow-probability=0.19.0=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.12.0+cu113=pypi_0
torchvision=0.13.0+cu113=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.12=pypi_0
webcolors=1.12=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.38.4=pypi_0
wrapt=1.14.1=pypi_0
xz=5.2.6=h5eee18b_0
zipp=3.12.0=pypi_0
zlib=1.2.13=h5eee18b_0
